requirements.txt
# Core LangChain & LangGraph
langgraph>=0.2.28
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.1
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0

# Groq Integration (Free LLM)
groq>=0.9.0
langchain-groq>=0.2.0

# GitHub Integration
PyGithub>=2.1.1

# Web Search
tavily-python>=0.5.0

# Vector Store & Embeddings (for RAG)
faiss-cpu>=1.8.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tiktoken>=0.7.0
pydantic>=2.8.2

# Frontend
streamlit>=1.31.0

# NEW: FastAPI for health check API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0


